Refactored Home.js into subcomponents. It was getting long and hard to maintain.
